1. add "make_windows.bat"  for build in Windows with  VisualStudio 2. fix some warnings

+++ src/file_logger.c
@@ -39,10 +39,10 @@ file_logger_handle file_logger_init(file_logger_cfg cfg)
 	char* log_folder_path_formatted = strreplace(cfg.log_folder_path, "\\", "/");
 	strlcpy(cfg.log_folder_path, log_folder_path_formatted, MAX_LOG_FOLDER_PATH_LEN);
 	free(log_folder_path_formatted);
-	int log_folder_path_len = strlen(cfg.log_folder_path);
+	size_t log_folder_path_len = strlen(cfg.log_folder_path);
 	if (cfg.log_folder_path[log_folder_path_len - 1] != '/')
 	{
-		int slash_location = (log_folder_path_len + 1) < MAX_LOG_FOLDER_PATH_LEN ?
+		size_t slash_location = (log_folder_path_len + 1) < MAX_LOG_FOLDER_PATH_LEN ?
 			log_folder_path_len : (log_folder_path_len - 1);
 		cfg.log_folder_path[slash_location] = '/';
 		cfg.log_folder_path[slash_location + 1] = '\0';
@@ -62,7 +62,7 @@ file_logger_handle file_logger_init(file_logger_cfg cfg)
 		return NULL;
 	}
 	handle->cfg = cfg;
-	handle->msg_queue = QueueHandler_create(cfg.max_log_queue_size, handle_log_queue_msg, handle);
+	handle->msg_queue = QueueHandler_create((uint32_t)cfg.max_log_queue_size, handle_log_queue_msg, handle);
 	if (NULL == handle->msg_queue)
 	{
 		free(handle);
@@ -147,7 +147,7 @@ static void handle_log_queue_msg(queue_msg_t* msg_p, void* user_data)
 	{
 		return;
 	}
-	int log_msg_len = strlen(msg_p->obj.data);
+	size_t log_msg_len = strlen(msg_p->obj.data);
 	if (log_msg_len <= 0)
 	{
 		return;
